+++ packages/apollo-gateway/src/index.ts
@@ -152,6 +152,15 @@ type RequestContext<TContext> = WithRequired<
   'document' | 'queryHash'
 >;
 
+// Local state to track whether particular UX-improving warning messages have
+// already been emitted.  This is particularly useful to prevent recurring
+// warnings of the same type in, e.g. repeating timers, which don't provide
+// additional value when they are repeated over and over during the life-time
+// of a server.
+type WarnedStates = {
+  remoteWithLocalConfig?: boolean;
+};
+
 export class ApolloGateway implements GraphQLService {
   public schema?: GraphQLSchema;
   protected serviceMap: DataSourceCache = Object.create(null);
@@ -164,6 +173,7 @@ export class ApolloGateway implements GraphQLService {
   private serviceDefinitions: ServiceDefinition[] = [];
   private compositionMetadata?: CompositionMetadata;
   private serviceSdlCache = new Map<string, string>();
+  private warnedStates: WarnedStates = Object.create(null);
 
   private fetcher: typeof fetch = fetcher.defaults({
     cacheManager: new HttpRequestCache(),
@@ -465,6 +475,38 @@ export class ApolloGateway implements GraphQLService {
   protected async loadServiceDefinitions(
     config: GatewayConfig,
   ): ReturnType<Experimental_UpdateServiceDefinitions> {
+    // This helper avoids the repetition of options in the two cases this method
+    // is invoked below. It is a helper, rather than an options object, since it
+    // depends on the presence of `this.engineConfig`, which is guarded against
+    // further down in this method in two separate places.
+    const getRemoteConfig = (engineConfig: GraphQLServiceEngineConfig) => {
+      return getServiceDefinitionsFromStorage({
+        graphId: engineConfig.graphId,
+        apiKeyHash: engineConfig.apiKeyHash,
+        graphVariant: engineConfig.graphVariant,
+        federationVersion:
+          (config as ManagedGatewayConfig).federationVersion || 1,
+        fetcher: this.fetcher,
+      });
+    };
+
+    if (isLocalConfig(config) || isRemoteConfig(config)) {
+      if (this.engineConfig && !this.warnedStates.remoteWithLocalConfig) {
+        // Only display this warning once per start-up.
+        this.warnedStates.remoteWithLocalConfig = true;
+        // This error helps avoid common misconfiguration.
+        // We don't await this because a local configuration should assume
+        // remote is unavailable for one reason or another.
+        getRemoteConfig(this.engineConfig).then(() => {
+          this.logger.warn(
+            "A local gateway service list is overriding an Apollo Graph " +
+            "Manager managed configuration.  To use the managed " +
+            "configuration, do not specifiy a service list locally.",
+          );
+        }).catch(() => {}); // Don't mind errors if managed config is missing.
+      }
+    }
+
     if (isLocalConfig(config)) {
       return { isNewSchema: false };
     }
@@ -490,13 +532,7 @@ export class ApolloGateway implements GraphQLService {
       );
     }
 
-    return getServiceDefinitionsFromStorage({
-      graphId: this.engineConfig.graphId,
-      apiKeyHash: this.engineConfig.apiKeyHash,
-      graphVariant: this.engineConfig.graphVariant,
-      federationVersion: config.federationVersion || 1,
-      fetcher: this.fetcher
-    });
+    return getRemoteConfig(this.engineConfig);
   }
 
   // XXX Nothing guarantees that the only errors thrown or returned in